Facilidad de guardar datos decimales o flotantes como unidad de medida en receta-productos

+++ database/migrations/2024_13_02_220053_create_receta_producto_table.php
@@ -17,7 +17,7 @@ class CreateRecetaProductoTable extends Migration
             $table->id('id_recetas_producto'); // Crea la columna id_recetas_producto como llave primaria
             $table->unsignedBigInteger('id_producto'); // Columna id_producto como FK
             $table->unsignedBigInteger('id_receta'); // Columna id_receta como FK
-            $table->integer('cantidad'); // Columna tiempo_preparacion (en minutos o unidades de tiempo)
+            $table->decimal('cantidad', 10, 2); // Cambiado a decimal con 2 decimales
             $table->boolean('estado')->default(true); // Columna estado (activo o inactivo)
             $table->timestamps(); // Crea las columnas created_at y updated_at
